refactor(chat): extract pure helpers to chat_helpers.py
Move side-effect-free utilities (validate_and_adjust_max_tokens, is_free_model,
mask_key, _to_thread, _get_auto/code_route_default_model) out of the 3236-line
chat.py into a focused module, re-imported for src.routes.chat.* patch-compat.
Zero behavior change (CM suite identical; stream_generator import intact).
